python-backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
numpy
pandas
scikit-learn
google-generativeai>=0.3.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
httpx
pydantic-settings==2.1.0
redis==5.0.1
celery==5.3.4
supabase==2.0.0
websockets>=11.0.3